+++ sos/policies/distros/debian.py
@@ -1,5 +1,14 @@
+# This file is part of the sos project: https://github.com/sosreport/sos
+#
+# This copyrighted material is made available to anyone wishing to use,
+# modify, copy, or redistribute it subject to the terms and conditions of
+# version 2 of the GNU General Public License.
+#
+# See the LICENSE file in the source distribution for further information.
+
 from sos.report.plugins import DebianPlugin
-from sos.policies import PackageManager, LinuxPolicy
+from sos.policies.distros import LinuxPolicy
+from sos.policies import PackageManager
 
 import os
