+++ sos/policies/distros/ubuntu.py
@@ -1,5 +1,13 @@
+# This file is part of the sos project: https://github.com/sosreport/sos
+#
+# This copyrighted material is made available to anyone wishing to use,
+# modify, copy, or redistribute it subject to the terms and conditions of
+# version 2 of the GNU General Public License.
+#
+# See the LICENSE file in the source distribution for further information.
+
 from sos.report.plugins import UbuntuPlugin
-from sos.policies.debian import DebianPolicy
+from sos.policies.distros.debian import DebianPolicy
 
 import os
